fix: merge backend scope files to satisfy pkg-file-count limit
Consolidate backend_scope.go into factory_build.go and backend_scope_test.go
into runtime_session_runtime_test.go so pkg/service stays within the 15-file cap.
