+++ pkg/service/factory_build.go
@@ -19,7 +19,9 @@ import (
 	"github.com/portpowered/infinite-you/pkg/apisurface"
 	"github.com/portpowered/infinite-you/pkg/cli/dashboardrender"
 	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
+	"github.com/portpowered/infinite-you/pkg/config/defaultpaths"
 	"github.com/portpowered/infinite-you/pkg/config/operatorconfig"
+	"github.com/portpowered/infinite-you/pkg/config/systemconfig"
 	"github.com/portpowered/infinite-you/pkg/factory"
 	factory_context "github.com/portpowered/infinite-you/pkg/factory/context"
 	factoryevents "github.com/portpowered/infinite-you/pkg/factory/events"
@@ -1804,4 +1806,55 @@ func modelHostDiagnostics(cfg *FactoryServiceConfig, logger *zap.Logger) modelho
 		diagnostics.Metrics = newModelHostMetricsRecorder(cfg.InvocationMetricsRecorder)
 	}
 	return diagnostics
+}
+
+func ensureServiceBackendScope(cfg *FactoryServiceConfig, logger *zap.Logger) error {
+	if cfg == nil {
+		return fmt.Errorf("factory service config is required to resolve backend scope")
+	}
+	if strings.TrimSpace(cfg.ReplayPath) != "" {
+		return nil
+	}
+	if strings.TrimSpace(cfg.BackendScopeID) != "" {
+		return nil
+	}
+
+	configPath, err := resolveSystemConfigPath(cfg)
+	if err != nil {
+		return err
+	}
+	resolved, err := systemconfig.EnsureLocalBackendScope(configPath)
+	if err != nil {
+		return err
+	}
+	cfg.BackendScopeID = resolved.BackendScopeID
+	if logger != nil {
+		logger.Info("resolved backend scope for local backend", zap.String("diagnostics", resolved.DiagnosticsLine()))
+	}
+	return nil
+}
+
+func resolveSystemConfigPath(cfg *FactoryServiceConfig) (string, error) {
+	if cfg != nil && strings.TrimSpace(cfg.SystemConfigPath) != "" {
+		return strings.TrimSpace(cfg.SystemConfigPath), nil
+	}
+	homeDir := ""
+	if cfg != nil {
+		homeDir = strings.TrimSpace(cfg.SystemConfigHomeDir)
+	}
+	if homeDir == "" {
+		var err error
+		homeDir, err = os.UserHomeDir()
+		if err != nil {
+			return "", fmt.Errorf("resolve home directory for backend scope system config: %w", err)
+		}
+	}
+	return defaultpaths.OperatorConfigPath(homeDir), nil
+}
+
+func serviceBackendScopeID(cfg *FactoryServiceConfig) string {
+	if cfg == nil {
+		return ""
+	}
+	return strings.TrimSpace(cfg.BackendScopeID)
 }
